requirements.txt
alembic==1.14.1
arrow==1.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
Mako==1.3.9
MarkupSafe==3.0.2
nltk==3.9.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
Redis-Sentinel-Url==1.0.1
regex==2024.11.6
requests==2.32.3
rq==2.1.0
rq-dashboard==0.8.2.2
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.38
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
